+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/controllers/__init__.py
@@ -5,15 +5,7 @@
 
 from __future__ import annotations
 
-# enable motion generation extension
-from omni.isaac.core.utils.extensions import enable_extension
+from .differential_ik import DifferentialIKController
+from .differential_ik_cfg import DifferentialIKControllerCfg
 
-# TODO: Maybe keep this extension enabled by default? -- Fix the app experience.
-enable_extension("omni.isaac.motion_generation")
-
-
-# get module libraries
-from .differential_inverse_kinematics import DifferentialInverseKinematics
-from .rmp_flow import RmpFlowController
-
-__all__ = ["DifferentialInverseKinematics", "RmpFlowController"]
+__all__ = ["DifferentialIKController", "DifferentialIKControllerCfg"]